+++ plugin-server/tests/main/ingestion-queues/run-ingestion-pipeline.test.ts
@@ -66,7 +66,7 @@ describe('workerTasks.runEventPipeline()', () => {
             String(teamId),
             event.distinct_id
         )
-        const groupStoreForDistinctId = new BatchWritingGroupStoreForDistinctIdBatch(hub.db, new Map())
+        const groupStoreForDistinctId = new BatchWritingGroupStoreForDistinctIdBatch(hub.db, new Map(), new Map())
         await expect(
             new EventPipelineRunner(
                 hub,
